pyproject.toml
[project]
name = "memento-teams"
version = "0.1.0"
description = "Memento-Teams: Multi-agent orchestrator with TUI"
requires-python = ">=3.12,<3.13"
dependencies = [
    "textual",
    "rich",
    "python-dotenv",
    "langchain-openai",
    "langchain",
    "langchain-core",
    "langchain-mcp-adapters",
    "langgraph",
    "fastmcp",
    "httpx",
    "aiohttp",
    "aiofiles",
    "anyio",
    "litellm",
    "tiktoken",
    "pyyaml",
    "google_search_results",
    "requests",
    "pydantic",
    "pydantic-settings",
    "sentence-transformers",
    "chromadb",
    "rank-bm25",
    "anthropic",
    "openai",
    "nltk",
    "jieba>=0.42.1",
    "crawl4ai",
    "watchdog",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]
include = [
    "main.py",
    "tui_app.py",
]
